Alban update
Work done to answer questions about diagrams. Maybe getInfo should be moved into views.py, I don't know.
I changed the argument for the @permission_classes() decorator to IsAuthenticated. I think this is the right way to 
make sure login is required, but I am not sure and I do not have a test user to try that with. I was working with AllowAny for the summer.